--- a/CHANGELOG-unreleased.md
+++ b/CHANGELOG-unreleased.md
@@ -10,7 +10,7 @@ the released changes.
 ## Unreleased
 ### Changed
 - Avoided unnecessary creation of `SkyCoord` objects in `AstrometryEquatorial` and `AstrometryEcliptic`.
-- Avoided unnecessary array slices in `SolarSystemShapiro`
+- Avoided unnecessary `TOA` table slices in `SolarSystemShapiro`
 ### Added
 ### Fixed
 ### Removed