--- a/src/pint/simulation.py
+++ b/src/pint/simulation.py
@@ -3,7 +3,7 @@
 
 from collections import OrderedDict
 from copy import deepcopy
-from typing import Optional, List, Union
+from typing import Optional, List, Tuple, Union
 import pathlib
 
 import astropy.units as u
@@ -566,7 +566,7 @@ def calculate_random_models(
     keep_models: bool = True,
     return_time: bool = False,
     params: str = "all",
-) -> (np.ndarray, Optional[list]):
+) -> Tuple[np.ndarray, Optional[list]]:
     """
     Calculates random models based on the covariance matrix of the `fitter` object.
 
@@ -695,7 +695,7 @@ def _get_freqs_and_times(
     ntoas: int,
     freqs: u.Quantity,
     multi_freqs_in_epoch: bool = True,
-) -> (Union[float, u.Quantity, time.Time], np.ndarray):
+) -> Tuple[Union[float, u.Quantity, time.Time], np.ndarray]:
     freqs = np.atleast_1d(freqs)
     assert (
         len(freqs.shape) == 1 and len(freqs) <= ntoas