Avoid creating `SkyCoord` objects during `Astrometry` computations. This was partly implemented in #1646 for models with proper motion. This PR does the same thing for cases without proper motion. This has a significant performance advantage.
The tests introduced by #1646 also cover these changes.

In `SolarSystemShapiro`, avoid doing expensive and unnecessary slices of the `TOAs` table.

Spun off from #1743, to make testing easier.