@abhisrkckl 2024-04-29
Performance
-----------------

Here is the profiling code:
```
from pint.models import get_model_and_toas
from pint.config import examplefile
from pint.residuals import Residuals

m, t = get_model_and_toas(examplefile("NGC6440E.par.good"), examplefile("NGC6440E.tim"))
res = Residuals(t, m)

%timeit res.calc_phase_resids()
```

Before: 13.3 ms ± 45 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

After: 6.01 ms ± 19.3 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)