@dlakaplan 2024-04-30
I think this looks good.  Is any explicit new testing required, or is the fact that all of the usual tests passed here sufficient (I suspect the latter since you haven't yet changed the precision)?

Any changes wrt skipping the SkyCoord creation that need tests?